Build history report; Refactor major components into files.
