+++ src/Libraries/Microsoft.Extensions.AI.Evaluation.Reporting/TypeScript/components/DiagnosticsContent.tsx
@@ -1,0 +1,31 @@
+import { DismissCircle16Regular, Warning16Regular, Info16Regular } from "@fluentui/react-icons";
+import { useStyles } from "./Styles";
+
+
+export const DiagnosticsContent = ({ diagnostics }: { diagnostics: EvaluationDiagnostic[]; }) => {
+    const classes = useStyles();
+
+    const errorDiagnostics = diagnostics.filter(d => d.severity === "error");
+    const warningDiagnostics = diagnostics.filter(d => d.severity === "warning");
+    const infoDiagnostics = diagnostics.filter(d => d.severity === "informational");
+
+    return (
+        <>
+            {errorDiagnostics.map((diag, index) => (
+                <div key={`error-${index}`} className={classes.failMessage}>
+                    <DismissCircle16Regular /> {diag.message}
+                </div>
+            ))}
+            {warningDiagnostics.map((diag, index) => (
+                <div key={`warning-${index}`} className={classes.warningMessage}>
+                    <Warning16Regular /> {diag.message}
+                </div>
+            ))}
+            {infoDiagnostics.map((diag, index) => (
+                <div key={`info-${index}`} className={classes.infoMessage}>
+                    <Info16Regular /> {diag.message}
+                </div>
+            ))}
+        </>
+    );
+};
