+++ src/Libraries/Microsoft.Extensions.AI.Evaluation.Reporting/TypeScript/components/ScoreNodeHistory.tsx
@@ -1,10 +1,80 @@
 import { Table, TableHeader, TableRow, TableHeaderCell, TableBody, TableCell, TableCellLayout, Button, mergeClasses } from "@fluentui/react-components";
 import { useReportContext } from "./ReportContext";
 import { useStyles } from "./Styles";
-import { ScoreNodeType } from "./Summary";
-import { PassFailBar } from "./PassFailBar";
-import { PassFailBadge } from "./ScenarioTree";
+import { ScoreNode, ScoreNodeType, ScoreSummary } from "./Summary";
+import { PassFailBarLabel, PassFailVerticalBar } from "./PassFailBar";
 import { DismissRegular } from "@fluentui/react-icons";
+
+type ScoreNodeHistoryData = {
+    executionName: string,
+    selected: boolean,
+    hasData: boolean,
+    iterationPass: number,
+    iterationTotal: number,
+    scenarioPass: number,
+    scenarioTotal: number,
+    previous?: ScoreNodeHistoryData,
+};
+
+const calculateScoreNodeHistoryData = (scoreSummary: ScoreSummary, selectedScoreNode: ScoreNode) : ScoreNodeHistoryData[] => {
+    
+    const data: ScoreNodeHistoryData[] = [];
+    
+    const executions = [...scoreSummary.executionHistory.keys()].reverse();
+    for (const executionName of executions) {
+        const historyNode = scoreSummary.nodesByKey.get(executionName)?.get(selectedScoreNode.nodeKey);
+        if (!historyNode) {
+            data.push({
+                executionName, 
+                selected: false,
+                hasData: false, 
+                iterationPass: 0, 
+                iterationTotal: 0, 
+                scenarioPass: 0, 
+                scenarioTotal: 0
+            });
+            continue;
+        }
+
+        let itCtPass, itCtFail, scCtPass, scCtFail;
+        switch (historyNode.nodeType) {
+            case ScoreNodeType.Group:
+                itCtPass = historyNode.numPassingIterations;
+                itCtFail = historyNode.numFailingIterations;
+                scCtPass = historyNode.numPassingScenarios;
+                scCtFail = historyNode.numFailingScenarios;
+                break;
+            case ScoreNodeType.Scenario:
+                itCtPass = historyNode.numPassingIterations;
+                itCtFail = historyNode.numFailingIterations;
+                scCtPass = historyNode.numPassingScenarios;
+                scCtFail = historyNode.numFailingScenarios;
+                break;
+            case ScoreNodeType.Iteration:
+                itCtPass = historyNode.failed ? 0 : 1;
+                itCtFail = historyNode.failed ? 1 : 0;
+                scCtPass = historyNode.failed ? 0 : 1;
+                scCtFail = historyNode.failed ? 1 : 0;
+                break;
+        }
+
+        const previous = data.length > 0 ? data[data.length - 1] : undefined;
+
+        data.push({
+            executionName,
+            selected: historyNode.executionName === selectedScoreNode.executionName,
+            hasData: true,
+            iterationPass: itCtPass,
+            iterationTotal: itCtPass + itCtFail,
+            scenarioPass: scCtPass,
+            scenarioTotal: scCtPass + scCtFail,
+            previous,
+        });
+    
+    }
+
+    return data;
+}
 
 export const ScoreNodeHistory = () => {
     const classes = useStyles();
@@ -21,71 +91,82 @@ export const ScoreNodeHistory = () => {
         return null;
     }
 
-    const executions = Array.from(scoreSummary.executionHistory.keys()).reverse();
-
-    const getHistoryContent = (executionName: string) => {
-        const historyNode = scoreSummary.nodesByKey.get(executionName)?.get(summaryNode.nodeKey);
-        if (!historyNode) {
-            return undefined;
-        }
-
-        let ctPass, ctFail;
-        switch (historyNode.nodeType) {
-            case ScoreNodeType.Group:
-                ctPass = historyNode.numPassingIterations;
-                ctFail = historyNode.numFailingIterations;
-                break;
-            case ScoreNodeType.Scenario:
-                ctPass = historyNode.numPassingIterations;
-                ctFail = historyNode.numFailingIterations;
-                break;
-            case ScoreNodeType.Iteration:
-                ctPass = historyNode.failed ? 0 : 1;
-                ctFail = historyNode.failed ? 1 : 0;
-                break;
-        }
-
-        return (<>
-            <PassFailBar pass={ctPass} total={ctPass + ctFail} width="24px" height="12px" />
-            <PassFailBadge pass={ctPass} total={ctPass + ctFail} />
-        </>);
-    }
-
     const latestExecutionStyle = mergeClasses(classes.verticalText, classes.currentExecutionForeground);
-    
+
+    const historyData = calculateScoreNodeHistoryData(scoreSummary, summaryNode);
+
     return (<div className={classes.section}>
         <div className={classes.sectionHeader}>
-            <Button icon={<DismissRegular/>} appearance="transparent" onClick={() => selectScenarioLevel(selectedScenarioLevel)}/>
+            <Button icon={<DismissRegular />} appearance="transparent" onClick={() => selectScenarioLevel(selectedScenarioLevel)} />
             <h3 className={classes.sectionHeaderText}>Pass/Fail Trends for {summaryNode.name}</h3>
         </div>
 
         <div className={classes.sectionContainer}>
-            <div className={classes.tableContainer}>
-                <Table>
-                    <TableHeader>
-                        <TableRow>
-                            {executions.map((execution) => (
-                                <TableHeaderCell key={execution}
-                                    className={execution == summaryNode.executionName ? classes.currentExecutionBackground : undefined}>
-                                    <div className={classes.executionHeaderCell}>
-                                        <span className={execution == summaryNode.executionName ? latestExecutionStyle : classes.verticalText}>{execution}</span>
-                                    </div>
-                                </TableHeaderCell>
-                            ))}
-                        </TableRow>
-                    </TableHeader>
-                    <TableBody>
-                        <TableRow>
-                            {executions.map((execution) => (
-                                <TableCell key={execution}
-                                    className={execution == summaryNode.executionName ? classes.currentExecutionBackground : undefined}>
-                                    {getHistoryContent(execution)}
-                                </TableCell>
-                            ))}
-                        </TableRow>
-                    </TableBody>
-                </Table>
-            </div>
+            <Table>
+                <TableHeader>
+                    <TableRow>
+                        {historyData.map((data) => (
+                            <TableHeaderCell key={data.executionName}
+                                className={data.selected ? classes.currentExecutionBackground : undefined}>
+                                <div className={classes.executionHeaderCell}>
+                                    <span className={data.selected ? latestExecutionStyle : classes.verticalText}>{data.executionName}</span>
+                                </div>
+                            </TableHeaderCell>
+                        ))}
+                        <TableHeaderCell className={classes.currentExecutionBackground}>
+                        </TableHeaderCell>
+                    </TableRow>
+                </TableHeader>
+                <TableBody>
+                    <TableRow>
+                        {historyData.map((data) => (
+                            <TableCell key={data.executionName}
+                                className={data.selected ? classes.currentExecutionBackground : undefined}>
+                                {data.hasData && <HistoryCellContent data={data} mode="scenario"/>}
+                            </TableCell>
+                        ))}
+                        <TableCell className={classes.currentExecutionBackground}>
+                            <TableCellLayout className={classes.historyMetricCell}>Scenario<br />Pass Rate</TableCellLayout>
+                        </TableCell>
+                    </TableRow>
+                    <TableRow>
+                        {historyData.map((data) => (
+                            <TableCell key={data.executionName}
+                                className={data.selected ? classes.currentExecutionBackground : undefined}>
+                                {data.hasData && <HistoryCellContent data={data} mode="iteration"/>}
+                            </TableCell>
+                        ))}
+                        <TableCell className={classes.currentExecutionBackground}>
+                            <TableCellLayout className={classes.historyMetricCell}>Iteration<br />Pass Rate</TableCellLayout>
+                        </TableCell>
+                    </TableRow>
+                </TableBody>
+            </Table>
         </div>
+    </div>);
+};
+
+const HistoryCellContent = ({ data, mode}: {data: ScoreNodeHistoryData, mode: "iteration" | "scenario"}) => {
+    const classes = useStyles();
+
+    let ctPass, ctTotal, prevCtPass, prevCtTotal;
+    switch (mode) {
+        case "iteration":
+            ctPass = data.iterationPass;
+            ctTotal = data.iterationTotal;
+            prevCtPass = data.previous?.iterationPass;
+            prevCtTotal = data.previous?.iterationTotal;
+            break;
+        case "scenario":
+            ctPass = data.scenarioPass;
+            ctTotal = data.scenarioTotal;
+            prevCtPass = data.previous?.scenarioPass;
+            prevCtTotal = data.previous?.scenarioTotal;
+            break;
+    }
+
+    return (<div className={classes.scenarioHistoryCell}>
+        <PassFailVerticalBar pass={ctPass} total={ctTotal} width="24px" height="48px" />
+        <PassFailBarLabel pass={ctPass} total={ctTotal} prevPass={prevCtPass} prevTotal={prevCtTotal}/>
     </div>);
 };
